backend/requirements.txt
fastapi
uvicorn
openai
pandas
numpy
scikit-learn
supabase
python-dotenv
